argument resolver improved and moved
