+++ todo/src/main/java/com/benchmark/todo/_core/auth/argumentresolver/CurrentUserArgumentResolver.java
@@ -1,4 +1,4 @@
-package com.benchmark.todo.user.argumentresolver;
+package com.benchmark.todo._core.auth.argumentresolver;
 
 import com.benchmark.todo.user.entity.User;
 import com.benchmark.todo.user.service.UserService;
@@ -6,12 +6,14 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
 @RequiredArgsConstructor
+@Component
 public class CurrentUserArgumentResolver implements HandlerMethodArgumentResolver {
 
   private final UserService userService;
@@ -26,8 +28,6 @@ public class CurrentUserArgumentResolver implements HandlerMethodArgumentResolve
       NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
 
     Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-    String username = authentication.getName();
-
-    return userService.fetchUser(username);
+    return (User) authentication.getPrincipal();
   }
 }
