+++ todo/src/main/java/com/benchmark/todo/user/controller/UserController.java
@@ -1,11 +1,10 @@
 package com.benchmark.todo.user.controller;
 
-import com.benchmark.todo.user.argumentresolver.CurrentUser;
+import com.benchmark.todo._core.auth.argumentresolver.CurrentUser;
 import com.benchmark.todo.user.dto.UserDTO;
 import com.benchmark.todo.user.entity.User;
 import com.benchmark.todo.user.service.UserService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
