+++ todo/src/main/java/com/benchmark/todo/user/entity/User.java
@@ -57,4 +57,15 @@ public class User implements UserDetails {
   public boolean isEnabled() {
     return true;
   }
+
+  @Override
+  public String toString() {
+    return "User{" +
+        "id=" + id +
+        ", username='" + username + '\'' +
+        ", password='" + password + '\'' +
+        ", name='" + name + '\'' +
+        ", todoList=" + todoList +
+        '}';
+  }
 }
